+++ src/main/java/com/fikky/repositories/ChapterLinkRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 public interface ChapterLinkRepository extends CrudRepository<ChapterLink, Integer> {
-    public List<ChapterLink> findByChildId(Integer childId);
-    public List<ChapterLink> findByParentId(Integer parentId);
+    List<ChapterLink> findByChildId(Integer childId);
+    List<ChapterLink> findByParentId(Integer parentId);
 }
